Few corrections
